@@ -1,11 +1,12 @@
 import MEGASdk
+import MEGASwift
 
 private let sdkDelegateQueue = DispatchQueue(label: "nz.mega.MEGASDKRepo.MEGASdkAdditions")
 private let sdkCompletedTransfersProcessingQueue = DispatchQueue(label: "nz.mega.MEGASDKRepo.completedTransfersProcessingQueue")
 
 public extension MEGASdk {
     /// Associates a `NSMutableArray` of completed transfers with every **instance** of `MEGASdk`
-    private static var completedTransfers = [ObjectIdentifier: NSMutableArray]()
+    private static let completedTransfers: Atomic<[ObjectIdentifier: NSMutableArray]> = .init(wrappedValue: [:])
 
     @objc var completedTransfers: NSMutableArray {
         sdkCompletedTransfersProcessingQueue.sync {
@@ -38,44 +39,44 @@ public extension MEGASdk {
         CurrentUserSource.shared.isLoggedIn
     }
     
-    @objc func removeMEGADelegateAsync(_ delegate: any MEGADelegate) {
-        Task.detached {
-            MEGASdk.sharedSdk.remove(delegate)
-        }
-    }
-    
-    @objc func removeMEGARequestDelegateAsync(_ delegate: any MEGARequestDelegate) {
-        Task.detached {
-            MEGASdk.sharedSdk.remove(delegate)
-        }
-    }
-    
-    @objc func addMEGAGlobalDelegateAsync(_ delegate: any MEGAGlobalDelegate, queueType: ListenerQueueType) {
-        sdkDelegateQueue.async { [weak self] in
-            self?.add(delegate, queueType: queueType)
-        }
-    }
-    
-    @objc func removeMEGAGlobalDelegateAsync(_ delegate: any MEGAGlobalDelegate) {
+    @objc func removeMEGADelegateAsync(_ delegate: any MEGADelegate & Sendable) {
         sdkDelegateQueue.async { [weak self] in
             self?.remove(delegate)
         }
     }
     
-    @objc func removeMEGATransferDelegateAsync(_ delegate: any MEGATransferDelegate) {
-        Task.detached {
-            MEGASdk.sharedSdk.remove(delegate)
+    @objc func removeMEGARequestDelegateAsync(_ delegate: any MEGARequestDelegate & Sendable) {
+        sdkDelegateQueue.async { [weak self] in
+            self?.add(delegate)
+        }
+    }
+    
+    @objc func addMEGAGlobalDelegateAsync(_ delegate: any MEGAGlobalDelegate & Sendable, queueType: ListenerQueueType) {
+        sdkDelegateQueue.async { [weak self] in
+            self?.add(delegate, queueType: queueType)
+        }
+    }
+    
+    @objc func removeMEGAGlobalDelegateAsync(_ delegate: any MEGAGlobalDelegate & Sendable) {
+        sdkDelegateQueue.async { [weak self] in
+            self?.remove(delegate)
+        }
+    }
+    
+    @objc func removeMEGATransferDelegateAsync(_ delegate: any MEGATransferDelegate & Sendable) {
+        sdkDelegateQueue.async { [weak self] in
+            self?.remove(delegate)
         }
     }
     
     private var privateCompletedTransfers: NSMutableArray {
         let key = ObjectIdentifier(self)
-        if let completedTransfers = MEGASdk.completedTransfers[key] {
+        if let completedTransfers = MEGASdk.completedTransfers.wrappedValue[key] {
             return completedTransfers
         }
         
         let completedTransfers = NSMutableArray()
-        MEGASdk.completedTransfers[key] = completedTransfers
+        MEGASdk.completedTransfers.mutate { $0[key] = completedTransfers }
         return completedTransfers
     }
 }
