@@ -1,6 +1,6 @@
 import MEGASdk
 
-extension MEGAError: @retroactive LocalizedError {
+extension MEGAError: @retroactive LocalizedError, @unchecked Sendable {
     public var errorDescription: String? {
         "\(name)(\(type.rawValue))"
     }
